musictest/requirements.txt
wheel
Django
Django-rest-framework
psycopg2-binary

